Restructured main()
I have separated routines in simple functions

Divided the display into 2 screens,
one for the ds18b20 and the other for the dht11 sensor.
This way I can use a bigger font.
Each screen also has new bitmaps.
In the future, screen switch will be managed by a button instead of a timer.